Separate Vrm10 Texture Enumerator
