+++ Assets/VRM10/Editor/ScriptedImporter/VrmScriptedImporterEditorGUI.cs
@@ -68,7 +68,7 @@ namespace UniVRM10
                 case Tabs.Materials:
                     if (m_parser != null)
                     {
-                        EditorMaterial.OnGUI(m_importer, m_parser, Vrm10MaterialImporter.EnumerateAllTexturesDistinct);
+                        EditorMaterial.OnGUI(m_importer, m_parser, Vrm10TextureEnumerator.EnumerateAllTexturesDistinct);
                     }
                     break;
