+++ Assets/VRM10/Runtime/IO/Texture.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 7fe27001b96146d3bbc53044ff91b536
+timeCreated: 1620374936
