+++ Assets/VRM10/Runtime/IO/Texture/Vrm10TextureEnumerator.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 7763318be4064ecba712dcc012dd1cb1
+timeCreated: 1620374912
