+++ Assets/VRM10/Runtime/IO/Vrm10MaterialImporter.cs
@@ -176,72 +176,5 @@ namespace UniVRM10
             }
             return param;
         }
-
-        /// <summary>
-        /// VRM-1 の thumbnail テクスチャー。gltf.textures ではなく gltf.images の参照であることに注意(sampler等の設定が無い)
-        /// 
-        /// MToonとは無関係だがとりあえずここに
-        /// </summary>
-        /// <param name="parser"></param>
-        /// <param name="vrm"></param>
-        /// <param name="value"></param>
-        /// <returns></returns>
-        public static bool TryGetMetaThumbnailTextureImportParam(GltfParser parser, UniGLTF.Extensions.VRMC_vrm.VRMC_vrm vrm, out (SubAssetKey, TextureImportParam) value)
-        {
-            if (vrm?.Meta == null || !vrm.Meta.ThumbnailImage.HasValue)
-            {
-                value = default;
-                return false;
-            }
-
-            // thumbnail
-            var imageIndex = vrm.Meta.ThumbnailImage.Value;
-            var gltfImage = parser.GLTF.images[imageIndex];
-            var name = TextureImportName.GetUnityObjectName(TextureImportTypes.sRGB, gltfImage.name, gltfImage.uri);
-
-            GetTextureBytesAsync getBytesAsync = () =>
-            {
-                var bytes = parser.GLTF.GetImageBytes(parser.Storage, imageIndex);
-                return Task.FromResult(GltfTextureImporter.ToArray(bytes));
-            };
-            var param = new TextureImportParam(name, gltfImage.GetExt(), gltfImage.uri, Vector2.zero, Vector2.one, default, TextureImportTypes.sRGB, default, default,
-               getBytesAsync, default, default,
-               default, default, default
-               );
-            var key = new SubAssetKey(typeof(Texture2D), name);
-            value = (key, param);
-            return true;
-        }
-
-        /// <summary>
-        /// glTF 全体で使うテクスチャーをユニークになるように列挙する
-        /// </summary>
-        /// <param name="parser"></param>
-        /// <returns></returns>
-        public static IEnumerable<(SubAssetKey, TextureImportParam)> EnumerateAllTexturesDistinct(GltfParser parser)
-        {
-            if (!UniGLTF.Extensions.VRMC_vrm.GltfDeserializer.TryGet(parser.GLTF.extensions, out UniGLTF.Extensions.VRMC_vrm.VRMC_vrm vrm))
-            {
-                throw new System.Exception("not vrm");
-            }
-
-            if (TryGetMetaThumbnailTextureImportParam(parser, vrm, out (SubAssetKey, TextureImportParam) thumbnail))
-            {
-                yield return thumbnail;
-            }
-
-            var used = new HashSet<SubAssetKey>();
-            for (int i = 0; i < parser.GLTF.materials.Count; ++i)
-            {
-                var param = GetMaterialParam(parser, i);
-                foreach (var (key, value) in param.EnumerateSubAssetKeyValue())
-                {
-                    if (used.Add(key))
-                    {
-                        yield return (key, value);
-                    }
-                }
-            }
-        }
     }
 }
